organised and made improvements
