+++ fooddelivery_app/App.js
@@ -1,22 +1,13 @@
 import React from "react";
-import { StyleSheet, Text, View } from "react-native";
-import { NavigationContainer } from "@react-navigation/native";
-import { createStackNavigator } from "@react-navigation/stack";
-import SignupScreen from "./lib/Screens/SignupScreen";
-import LoginScreen from "./lib/Screens/LoginScreen";
-import SelectSizeScreen from "./lib/Screens/SelectSizeScreen";
-
-const Stack = createStackNavigator();
+import {
+  StyleSheet,
+  Text,
+  View
+} from "react-native";
+import MainNavigator from './lib/navigation/MainNavigator';
 
 export default function App() {
-  return (
-    <NavigationContainer>
-      <Stack.Navigator screenOptions={{ headerShown: false }}>
-        <Stack.Screen name="LoginPage" component={LoginScreen} />
-        <Stack.Screen name="SignupScreen" component={SignupScreen} />
-      </Stack.Navigator>
-    </NavigationContainer>
-  );
+  return ( < MainNavigator / > );
 }
 
 const styles = StyleSheet.create({
