+++ fooddelivery_app/lib/navigation/MainNavigator.js
@@ -1,0 +1,24 @@
+import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
+import { createStackNavigator } from "@react-navigation/stack";
+import SignupScreen from "../Screens/SignupScreen";
+import LoginScreen from "../Screens/LoginScreen";
+import SelectSizeScreen from "../Screens/SelectSizeScreen";
+
+function MainNavigator() {
+  const Stack = createStackNavigator();
+  return (
+    <NavigationContainer>
+      <Stack.Navigator
+        screenOptions={{
+          headerShown: false
+        }}
+      >
+        <Stack.Screen name="LoginPage" component={LoginScreen} />
+        <Stack.Screen name="SignupScreen" component={SignupScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
+
+export default MainNavigator;
